+++ hapi/src/main/java/com/hedera/hapi/util/UnknownHederaFunctionality.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hedera.node.app.spi;
+package com.hedera.hapi.util;
 
 import com.hedera.hapi.node.base.HederaFunctionality;
 import com.hedera.hapi.node.transaction.Query;
