+++ hapi/src/main/java/module-info.java
@@ -53,6 +53,7 @@ module com.hedera.node.hapi {
     exports com.hedera.services.stream.proto;
     exports com.hederahashgraph.api.proto.java;
     exports com.hederahashgraph.service.proto.java;
+    exports com.hedera.hapi.util;
 
     requires transitive com.google.common;
     requires transitive com.google.protobuf;
