+++ hedera-node/hedera-app-spi/src/main/java/com/hedera/node/app/spi/state/Schema.java
@@ -16,7 +16,7 @@
 
 package com.hedera.node.app.spi.state;
 
-import static com.hedera.node.app.spi.HapiUtils.SEMANTIC_VERSION_COMPARATOR;
+import static com.hedera.hapi.util.HapiUtils.SEMANTIC_VERSION_COMPARATOR;
 
 import com.hedera.hapi.node.base.SemanticVersion;
 import com.hedera.pbj.runtime.Codec;
