+++ hedera-node/hedera-app-spi/src/test/java/com/hedera/node/app/spi/HapiUtilsTest.java
@@ -16,7 +16,7 @@
 
 package com.hedera.node.app.spi;
 
-import static com.hedera.node.app.spi.HapiUtils.asTimestamp;
+import static com.hedera.hapi.util.HapiUtils.asTimestamp;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.params.provider.Arguments.of;
 
@@ -25,6 +25,7 @@ import com.hedera.hapi.node.base.Key;
 import com.hedera.hapi.node.base.KeyList;
 import com.hedera.hapi.node.base.ThresholdKey;
 import com.hedera.hapi.node.base.Timestamp;
+import com.hedera.hapi.util.HapiUtils;
 import com.hedera.pbj.runtime.io.buffer.Bytes;
 import com.swirlds.platform.test.fixtures.state.TestBase;
 import edu.umd.cs.findbugs.annotations.NonNull;
